Reworked `lib` to solve auto/eager-loading issues
* Classes/modules from `lib` should be loaded explicitly during init
* The rest moved to `app/lib`, as recommended by Rails community
