+++ app/lib/errors/backend/authorization_error.rb
