+++ app/lib/errors/backend/entity_not_found_error.rb
