+++ config/application.rb
@@ -31,9 +31,6 @@ module ROCCIServer
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
-    # Autoload stuff from lib/
-    config.autoload_paths << Rails.root.join('lib')
-
     # Only loads a smaller set of middleware suitable for API only apps.
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
