+++ config/initializers/00_extensions.rb
@@ -1,0 +1,2 @@
+# Load extensions from Rails.root/lib/ext
+Dir[Rails.root.join('lib', 'ext', '*.rb')].each { |file| require file.gsub('.rb', '') }
