+++ config/initializers/01_rack_middleware.rb
@@ -1,0 +1,2 @@
+# Load extensions from Rails.root/lib/rack
+Dir[Rails.root.join('lib', 'rack', '*.rb')].each { |file| require file.gsub('.rb', '') }
