+++ config/routes.rb
@@ -4,14 +4,14 @@ Rails.application.routes.draw do
   ####################################################
   get '/',
       to: 'entity#list',
-      constraints: RoutingConstraints.build(%i[non_legacy])
+      constraints: Ext::RoutingConstraints.build(%i[non_legacy])
   get '/',
       to: 'entity#locations',
-      constraints: RoutingConstraints.build(%i[legacy])
+      constraints: Ext::RoutingConstraints.build(%i[legacy])
 
   post '/',
        to: 'entity#execute_all',
-       constraints: RoutingConstraints.build(%i[action])
+       constraints: Ext::RoutingConstraints.build(%i[action])
 
   delete '/', to: 'entity#delete_all'
 
@@ -19,25 +19,25 @@ Rails.application.routes.draw do
 
   get '/mixin/:parent/:term/',
       to: 'entity#scoped_list',
-      constraints: RoutingConstraints.build(%i[non_legacy mixin])
+      constraints: Ext::RoutingConstraints.build(%i[non_legacy mixin])
   get '/mixin/:parent/:term/',
       to: 'entity#scoped_locations',
-      constraints: RoutingConstraints.build(%i[legacy mixin])
+      constraints: Ext::RoutingConstraints.build(%i[legacy mixin])
 
   post '/mixin/:parent/:term/',
        to: 'entity#scoped_execute_all',
-       constraints: RoutingConstraints.build(%i[action mixin])
+       constraints: Ext::RoutingConstraints.build(%i[action mixin])
   post '/mixin/:parent/:term/',
        to: 'entity#assign_mixin',
-       constraints: RoutingConstraints.build(%i[non_action mixin])
+       constraints: Ext::RoutingConstraints.build(%i[non_action mixin])
 
   put '/mixin/:parent/:term/',
       to: 'entity#update_mixin',
-      constraints: RoutingConstraints.build(%i[mixin])
+      constraints: Ext::RoutingConstraints.build(%i[mixin])
 
   delete '/mixin/:parent/:term/',
          to: 'entity#scoped_delete_all',
-         constraints: RoutingConstraints.build(%i[mixin])
+         constraints: Ext::RoutingConstraints.build(%i[mixin])
 
   ####################################################
   ## Query Interface Routes
@@ -66,38 +66,38 @@ Rails.application.routes.draw do
   ####################################################
   get '/:resource/:id',
       to: 'resource#show',
-      constraints: RoutingConstraints.build(%i[resource])
+      constraints: Ext::RoutingConstraints.build(%i[resource])
   get '/:resource/',
       to: 'resource#list',
-      constraints: RoutingConstraints.build(%i[non_legacy resource])
+      constraints: Ext::RoutingConstraints.build(%i[non_legacy resource])
   get '/:resource/',
       to: 'resource#locations',
-      constraints: RoutingConstraints.build(%i[legacy resource])
+      constraints: Ext::RoutingConstraints.build(%i[legacy resource])
 
   post '/:resource/:id',
        to: 'resource#execute',
-       constraints: RoutingConstraints.build(%i[action resource])
+       constraints: Ext::RoutingConstraints.build(%i[action resource])
   post '/:resource/:id',
        to: 'resource#partial_update',
-       constraints: RoutingConstraints.build(%i[non_action resource])
+       constraints: Ext::RoutingConstraints.build(%i[non_action resource])
   post '/:resource/',
        to: 'resource#execute_all',
-       constraints: RoutingConstraints.build(%i[action resource])
+       constraints: Ext::RoutingConstraints.build(%i[action resource])
   post '/:resource/',
        to: 'resource#create',
-       constraints: RoutingConstraints.build(%i[non_action resource])
+       constraints: Ext::RoutingConstraints.build(%i[non_action resource])
 
   put '/:resource/:id',
       to: 'resource#update',
-      constraints: RoutingConstraints.build(%i[resource])
+      constraints: Ext::RoutingConstraints.build(%i[resource])
   # put '/:resource/' is undefined in GFD-P-R.185
 
   delete '/:resource/:id',
          to: 'resource#delete',
-         constraints: RoutingConstraints.build(%i[resource])
+         constraints: Ext::RoutingConstraints.build(%i[resource])
   delete '/:resource/',
          to: 'resource#delete_all',
-         constraints: RoutingConstraints.build(%i[resource])
+         constraints: Ext::RoutingConstraints.build(%i[resource])
 
   ####################################################
   ## Occi::Core::Link Routes
@@ -106,36 +106,36 @@ Rails.application.routes.draw do
   ####################################################
   get '/link/:link/:id',
       to: 'link#show',
-      constraints: RoutingConstraints.build(%i[link])
+      constraints: Ext::RoutingConstraints.build(%i[link])
   get '/link/:link/',
       to: 'link#list',
-      constraints: RoutingConstraints.build(%i[non_legacy link])
+      constraints: Ext::RoutingConstraints.build(%i[non_legacy link])
   get '/link/:link/',
       to: 'link#locations',
-      constraints: RoutingConstraints.build(%i[legacy link])
+      constraints: Ext::RoutingConstraints.build(%i[legacy link])
 
   post '/link/:link/:id',
        to: 'link#execute',
-       constraints: RoutingConstraints.build(%i[action link])
+       constraints: Ext::RoutingConstraints.build(%i[action link])
   post '/link/:link/:id',
        to: 'link#partial_update',
-       constraints: RoutingConstraints.build(%i[non_action link])
+       constraints: Ext::RoutingConstraints.build(%i[non_action link])
   post '/link/:link/',
        to: 'link#execute_all',
-       constraints: RoutingConstraints.build(%i[action link])
+       constraints: Ext::RoutingConstraints.build(%i[action link])
   post '/link/:link/',
        to: 'link#create',
-       constraints: RoutingConstraints.build(%i[non_action link])
+       constraints: Ext::RoutingConstraints.build(%i[non_action link])
 
   put '/link/:link/:id',
       to: 'link#update',
-      constraints: RoutingConstraints.build(%i[link])
+      constraints: Ext::RoutingConstraints.build(%i[link])
   # put '/link/:link/' is undefined in GFD-P-R.185
 
   delete '/link/:link/:id',
          to: 'link#delete',
-         constraints: RoutingConstraints.build(%i[link])
+         constraints: Ext::RoutingConstraints.build(%i[link])
   delete '/link/:link/',
          to: 'link#delete_all',
-         constraints: RoutingConstraints.build(%i[link])
+         constraints: Ext::RoutingConstraints.build(%i[link])
 end
